requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
passlib==1.7.4
pydantic==2.7.1
redis==5.0.1
python-jose[cryptography]==3.3.0
websockets==12.0
openai==1.30.0
